requirements.txt
streamlit==1.41.1
pyserial==3.5
